Add personal task deletion and view

+++ app/db/repository/groups.py
@@ -38,12 +38,10 @@ def find_admin_in_group_db(db, group_id: int):
 
 
 def delete_user_in_group_db(db, user_id: int, group_id: int):
-    query = db.query(UserInGroupDB)
-    query = query.filter(UserInGroupDB.user_id == user_id)
-    query = query.filter(UserInGroupDB.group_id == group_id)
-    user_in_group_db = query.first()
-    db.delete(user_in_group_db)
-    db.commit()
+    user_in_group_db = find_user_in_group_db(db, user_id, group_id)
+    if user_in_group_db:
+        db.delete(user_in_group_db)
+        db.commit()
 
 
 def get_user_groups_from_db(db, username: str):
